+++ wgsl_to_wgpu/src/snapshots/write_all_structs_bytemuck_input_layout_validation.snap.rs
@@ -79,10 +79,10 @@ pub fn vertex_state<'a, const N: usize>(
         },
     }
 }
-pub fn main_entry(input0: wgpu::VertexStepMode) -> VertexEntry<1> {
+pub fn main_entry(input_step_mode: wgpu::VertexStepMode) -> VertexEntry<1> {
     VertexEntry {
         entry_point: ENTRY_MAIN,
-        buffers: [Input0::vertex_buffer_layout(input0)],
+        buffers: [Input0::vertex_buffer_layout(input_step_mode)],
         constants: Default::default(),
     }
 }